refactor: move update_tg_ct & set_next_recruitment to DB manager
